fixed publishing issue with cluster environment
